Remove Upgrade Hanlder v7
