+++ app/upgrade/v8/upgrade.go
@@ -1,0 +1,31 @@
+package v8
+
+import (
+	"context"
+
+	store "cosmossdk.io/store/types"
+	upgradetypes "cosmossdk.io/x/upgrade/types"
+	"github.com/cosmos/cosmos-sdk/types/module"
+
+	"github.com/tokenize-x/tx-chain/v7/app/upgrade"
+)
+
+// Name defines the upgrade name.
+const Name = "v8"
+
+// New makes an upgrade handler for v8 upgrade.
+func New(
+	mm *module.Manager,
+	configurator module.Configurator,
+) upgrade.Upgrade {
+	return upgrade.Upgrade{
+		Name: Name,
+		StoreUpgrades: store.StoreUpgrades{
+			Added:   []string{},
+			Deleted: []string{},
+		},
+		Upgrade: func(ctx context.Context, _ upgradetypes.Plan, vm module.VersionMap) (module.VersionMap, error) {
+			return mm.RunMigrations(ctx, configurator, vm)
+		},
+	}
+}
