+++ integration-tests/upgrade/upgrade_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	appupgradev7 "github.com/tokenize-x/tx-chain/v7/app/upgrade/v7"
+	appupgradev8 "github.com/tokenize-x/tx-chain/v7/app/upgrade/v8"
 	integrationtests "github.com/tokenize-x/tx-chain/v7/integration-tests"
 	"github.com/tokenize-x/tx-chain/v7/testutil/integration"
 	"github.com/tokenize-x/tx-tools/pkg/retry"
@@ -43,23 +43,21 @@ func TestUpgrade(t *testing.T) {
 	infoRes, err := tmQueryClient.GetNodeInfo(ctx, &cmtservice.GetNodeInfoRequest{})
 	requireT.NoError(err)
 
-	if strings.HasPrefix(infoRes.ApplicationVersion.Version, "v6.") {
-		upgradeV6ToV7(t)
+	if strings.HasPrefix(infoRes.ApplicationVersion.Version, "v7.") {
+		upgradeV7ToV8(t)
 		return
 	}
 	requireT.Failf("not supported txd version", "version: %s", infoRes.ApplicationVersion.Version)
 }
 
-func upgradeV6ToV7(t *testing.T) {
-	tests := []upgradeTest{
-		&pseMigrationTest{},
-	}
+func upgradeV7ToV8(t *testing.T) {
+	tests := []upgradeTest{}
 
 	for _, test := range tests {
 		test.Before(t)
 	}
 
-	runUpgrade(t, appupgradev7.Name, upgradeDelayInBlocks)
+	runUpgrade(t, appupgradev8.Name, upgradeDelayInBlocks)
 
 	for _, test := range tests {
 		test.After(t)
